requirements.txt
websocket-client
openai
pyaudio
keyboard
python-dotenv
gpiozero
rpi.gpio
pigpio
spidev
lgpio
git+https://github.com/larsimmisch/pyalsaaudio.git